Update patch set 2
Patch Set 2:

Robert, to address the comments from Irfan on change https://review.source.android.com/#change,22644 I would need to change the interface to pass the result of the API call to the callee. Could you please remove the approved flag to make sure the change doesn't get merged yet or else I would have to create a new change to address that (which is not desirable)?

Patch-set: 2