Fix ACTION_CALL_EMERGENCY code path through OutgoingCallBroadcaster
InCallScreen/CallController overhaul: One place in OutgoingCallBroadcaster
was still launching the InCallScreen with a CALL intent, which would crash
the phone app if you tried to dial 911 :-(
The fix is to use CallController.placeCall() instead.

TESTED (crespo):
  - regular outgoing calls
  - directly dial 911
  - dial 911 via voice dialer

Also double-checked that nobody else is manually creating intents pointing
to InCallScreen.class, except for places that are already flagged with a
"TODO(InCallScreen redesign)" comment.

Bug: 4444071
Bug: 4194458

Change-Id: I518fbf71302564bfb31d1c821464961823f1ed8b