Update patch set 2
Patch Set 2:

The commit message is a summary of all four UICC changes, not only of this change. The emulator extension is optional, but the remaining three changes depend on each other. The UICC extension should build fine without the Open Mobile API changes.

The method openIccLogicalChannel(String AID) returns the session ID for a logical channel as defined in 3GPP TS 27.007. The session ID is defined as a 32-bit unsigned number. The value zero can be seen as invalid. The caller of the method openIccLogicalChannel must be able to distinguish between several error types. Currently we use the getLastError() method to get that information. 

Do you prefer a new class (IccOpenLogicalChannelResult) with session ID and error code as members?

Patch-set: 2