Set FLAG_AUTO_CANCEL on missed call notification
When the call log was converted from an activity to a fragment, we
accidentally broke the code that clears the "missed call" notification
upon viewing the call log.

That's slightly nontrivial to fix (see bug 4521535), so as a short-term
workaround let's instead set FLAG_AUTO_CANCEL on the initial notification.
This is only a partial fix, but at least the notification will be cleared
in the most common use case (i.e. selecting the missed call notification
from the windowshade).

Bug: 4521535
Change-Id: I8dddbf88f9b0728f0b623d6c67d02a577223b8f7