Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Looks good to me, but I'd like Daisuke to confirm too.

By the way, can you confirm that nothing bad happens if you launch this
activity with ACTION_VIEW but you have zero SIM contacts?  (The code in
ADNList.java seems to be careful about this case, but it would still be
nice to actually test that.)

Patch-set: 1