Only show geoDescription for incoming numbers
It doesn't really make sense to show the geoDescription for outgoing
numbers (i.e. numbers you manually dialed using the dialpad), since you
presumably have *some* knowledge about the number already.

So only display it for incoming calls.

(Note that you need a Connection, not a Call, to tell if it's incoming
or not.  Luckily we already have a Connection object at all the points
where we call updateDisplayForPerson(), so we now pass that as a
parameter too.)

Bug: 4595580
Change-Id: I2ea4e4213f8a4347082a3db949c31de4c25372ab