Reset the "respond via SMS" popup for new incoming calls
Fix a case where the "respond via SMS" popup could cause the incoming-call
UI to be stuck in a broken state for the *next* incoming call.

The fix: When a new incoming call arrives, check to see if there's still a
RespondViaSmsPopup instance around since last time.  If so, make sure it's
not still active!

The details:
  - RespondViaSms.showRespondViaSmsPopup() now returns the PopupMenu
    instance (so the InCallScreen can hold onto it from one call to the
    next).
  - InCallScreen now needs to listen for "new ringing connection" events
    in addition to regular "incoming ring" events.

TESTED the following sequence:
  (1) get an incoming call
  (2) drag upward to bring up the respond-via-SMS popop
  (3) wait till the incoming call goes to voicemail
  (4) get a 2nd incoming call
   ==> Before this change: Popup is still visible when 2nd call starts ringing(!)
       Now: Normal incoming-call UI appears.

Bug: 4598484
Change-Id: I9ead646439cad8e9f323e8983f144dba4e5916f6