Temp fix for onDismiss() from the "respond via SMS" popup
Temporarily disable the restartRinger() call when the "respond via SMS"
popup gets dismissed.

We're only supposed to start the ringer if the user manually dismisses the
popup, like by tapping outside it or pressing Back.  But the onDismiss()
callback gets called *any* time the popup is dismissed, even by the user
selecting one of the choices!

So as a quick fix for now, just comment out the restartRinger() call.  The
real fix is that we need a separate onCancel() callback from the popup
(see bug 4818050), and the current code in onDismiss() should be moved to
onCancel() instead.

Bug: 4598484
Change-Id: I4a12575589211f6c71de510eeec447560e856507