Rename RespondViaSms to RespondViaSmsManager, and make it instantiatable
Originally the RespondViaSms class was just a collection of static utility
methods.  But it now needs to maintain some persistent state (namely the
user's custom "canned responses") so it makes more sense for it to
actually be a real instance owned by the InCallScreen.

Note this is NOT a singleton; each InCallScreen instance has its own
RespondViaSmsManager instance.  (If RespondViaSmsManager were a singleton,
it would still need to point back to the active InCallScreen, and that
could lead to a memory leak if the ActivityManager were to destroy and
recreate the InCallScreen.)

Change-Id: Icb4d6cf4211fc1deed808408f31bd3854e183d1a