In-call UI: Replace Bluetooth / Speaker buttons with a 3-way popup
To reduce the number of onscreen buttons in the in-call UI, the new design
calls for a single "Audio mode" button that brings up a PopupMenu to
switch between Speaker / Handset earpiece / Bluetooth.

This replaces the two separate toggle buttons we had previously (one for
Speaker, one for Bluetooth.)

This change adds the popup menu, and adds InCallScreen.switchInCallAudio()
to implement the 3-way switch.  We don't have assets for the new buttons
yet, so for now the "Audio mode" button just looks like one of the old
toggle buttons (even though it doesn't actually toggle.)

Still missing:
  - Icons in the PopupMenu
  - When the PopupMenu comes up, the current audio mode should be
    highlighted initially
  - New assets for all the in-call buttons

Bug: 4965878

Change-Id: I390da4b3e26e5af86efe284038a6961a994a90f4