Fix cosmetic glitch with transient window when launching the InCallScreen
When making an outgoing call, before the in-call UI comes up there's
currently a brief flash of an intermediate screen with just a header
labeled "Phone".

That's the transient window that the framework creates to cover up the
screen transition, and it looks bad because that window doesn't match the
actual configuration of the resulting activity.

The fix is to make sure that the entire sequence of activities for an
outgoing call(*) ALL specify themes that tell the framework there's no
action bar.

(*) The exact details:

  Here's the sequence of 3 separate activities that get launched when you
  make an outgoing call:

  (1) OutgoingCallBroadcaster: an invisible activity that runs a hook that
      lets 3rd party apps intercept or rewrite the outgoing call, and then
      launches:

  (2) SipCallOptionHandler: an activity that *might* bring up a dialog if
      the user needs to choose between making this call via PSTN or SIP.
      Usually we don't need a dialog at all, so we just launch:

  (3) InCallScreen: the actual in-call UI

Currently in the manifest:
   (1) has no theme at all
   (2) has android:theme="@android:style/Theme.Holo.DialogWhenLarge"
   (3) has android:theme="@android:style/Theme.NoTitleBar"

The problem was activity (2).  It needs to specify
   android:theme="@android:style/Theme.Holo.NoActionBar"
to be clear that it doesn't have an action bar.
(The "Theme.Holo.DialogWhenLarge" is not needed since (a) this activity
launches a dialog explicitly if necessary, and (b) the activity is
phone-only anyway.)

Also, since activity (1) never shows any UI of its own, it's good style
for it to explicitly set android:theme="@android:style/Theme.NoDisplay",
so this change does that too.

TESTED:
  - outgoing call: glitch is gone
  - incoming call: no visible change
  - "Audio mode" popup: still looks the same as before
  - "Respond via SMS" dialog: still looks the same as before
  - "PSTN or SIP" dialog from SipCallOptionHandler: still looks the same as before

Bug: 5039564
Change-Id: I313f57c5d968685778513020181c1576dc07ad51