Do not trigger mode switch to default if LTE only mode is used.
If LTE only mode is used and Network Mode UI is opened,
the default network mode will be set and GSMPhone will be created
in the next reboot. And this was causing continuous phone process
crash with CDMA properties.

This patch is to handle LTE only mode (which is defined, but not
supported by UI) gracefully.

Bug: 5070048
Change-Id: I0e90ec119a4c6b248d104292dd082ce964bc315c