get LTE Data Service related code reconnected with the pref
Oddly, even though the code associated with the data service pref
was left behind when I moved the pref to another pref xml in
https://android-git.corp.google.com/g/114571, it continued to work
correctly in honeycomb-LTE, and still works fine today. But something
happened to finally sever the connection between then and now,
and so I'm now moving the code to the correct place. (Which I should have
done in the first place!)

bug: 5045551
Change-Id: I26847c08a2fac6026e5181f72a8f2b03453906b9