"Respond via SMS": after responding, reject the incoming call immediately.
Once the user responds by SMS, that means they're done dealing with the
incoming call, so there's no reason to keep the call around.  (And it's
especially confusing for the "incoming call" icon in the status bar to
still be visible.)

So now we reject the incoming call as soon as the user selects one of the
choices from the popup (either one of the canned responses, or "Custom
message".)

Also fix a crash that could happen if the ringing call goes away at the
exact moment we're bringing up the popup menu.

Bug: 5042540
Change-Id: I4d7ebcf278abe708e86dbeec89a6a0912d7146ae