Clear missed calls when clearing their notification.
Currently, when the missed call notification is dismissed, no action is
taken. This means the phone still believes those missed calls need to be
shown to the user and on reboot it will add them back to the
notification.

This change adds a deleteIntent, which will be triggered when the
notification is dismissed. The service answering to the intent will in
term update the call log and clear the notification count.

Bug: 2142718
Change-Id: I0ee2270f5aca9380a808d7bd3d7e208be838581c