Reposition "Calling via ..." provider badge to not overlap the call banner
Currently, the badge that shows "provider info" (like "Calling via Google
Voice", for example) is drawn at the very top of the screen, and covers up
the info about the actual person / number you're calling :-(

In the short term, at least, the fix is to simply reposition the badge to
not collide with the call banner.  Post-ICS, though, we hope to integrate
the badge more smoothly into the call banner rather than using an overlay
at all (bug 5045539).

TESTED: Using google voice, verified that the badge is repositioned
  vertically more toward the middle of the screen, and doesn't obscure
  either the "call banner" area or any onscreen buttons.

Bug: 5045539
Change-Id: I4ec9f27b2edaa468d4c39515a94310169ffbf351