Fix SCO audio APIs.
The original meaning of the virtual call variable was to
check if we have sent the CIEV commands to start the virtual call.
However, this meaning changed later on.

In accordance with the original use, the virtual call started variable
was set only when we were able to send CIEV commands. We can send those
commands only to handsfree devices and not to devices which support
just the headset profile. Some devices don't have telephony
support and hence we only support headset profile for such devices.

9760482b introduced the bug where we were checking this variable
during terminate but not setting it during initiation.

Change-Id: I67b870841cff2fd8d9765b4ae4ed8fb49a1d1da4