Incoming call: add (faster) auto-repeat for "ping" animation
Previously, the ping animation was driven by RING events from the
telephony network, which come about 3 seconds apart.  That's so slow that
it's easy to miss the animation entirely, though.

So now, just manually auto-repeat the pings at a faster rate.

(I left the old RING event plumbing in place (but disabled) in case we
ever want something like that again in the future, though.)

TESTED:
  - incoming call: verified the "ping" happens more frequently
  - answer the call: verified MultiWaveView goes away correctly
  - reject the call: verified MultiWaveView goes away correctly
  - ignore the call: verified MultiWaveView goes away correctly

Bug: 5149690

Change-Id: Ie7c9fc6ea639505b66722be91e75dc29a7761ee9