In-call UI buttons: use new "toggle" style
Make "Dialpad", "Hold", and "Mute" be toggle buttons again (like they were
in Gingerbread).

The new asset ic_active_state_dialer_holo_dark is the "checked" indicator:
it's a blue bar along the bottom edge of the button.

Unfortunately there's no built-in framework widget that does exactly what
we want here, since (1) we want a single icon as the foreground image, and
(2) we need the concept of "checked" vs. "unchecked".  (1) comes from
ImageButton, and (2) comes from the CompoundButton base class, but those
are separate branches of the class hierarchy and there's no way to have
both :-(

So instead, implement these as ToggleButtons, and use an explicit
<layer-list> as the ToggleButton background, to render the button icon on
top and the "checked" indicator underneath.

Bug: 5117705

Change-Id: Ib2cf802062100d0ce24b2b9bf14f20773ed4343f