Don't hide in-call touch controls during "Call ended" state
Previously, we would completely hide the InCallTouchUi widget (the button
cluster at the bottom of the screen) during the "Call ended" state that's
shown briefly after a call ends.

But it's distracting for the buttons to disappear completely, since the
*rest* of the screen stays mostly the same as it was while in-call.  So
allow the buttons to stay visible in this state.

Also, add a couple of minor tweaks to the InCallControlState logic to make
sure the buttons are disabled if the phone is completely idle,

TESTED:
  - "Call ended" state after hanging up locally
  - "Call ended" state after having the other end hang up
  - Try pressing the various in-call buttons during the "Call ended"
    state; no harmful effects (since they're all disabled)

Bug: 5105773

Change-Id: I8373175a3fbc29a4a8f00977ce1c80ffc245980f