"Respond via SMS": disable the feature in cases we know it won't work
In the few cases where we know a priori that the "Respond via SMS" feature
isn't available for an incoming call, we now totally disable the "SMS"
drag target of the incoming call widget.

Those (relatively rare) cases are:
  - incoming number is bogus or blank
  - incoming "number" is a SIP address
  - incoming "call presentation" says that the number is blocked

Implementation note: There's no convenient way to dynamically enable /
disable one specific target of the MultiWaveView widget, so we just use
two sets of resource arrays: one set for the usual 3-way choice, and one
for the 2-way choice when you can't SMS.  Each time we bring up the
MultiWaveView widget we update its resources to use the correct arrays.

TESTED:
  - Normal incoming calls
      ==> "Respond via SMS" still available
  - Manually hacked allowRespondViaSmsForCall() to return false:
      ==> Incoming call widget only shows "Answer" and "Decline" choices
      ==> Dragging upward and letting go has no effect
  - I didn't test allowRespondViaSmsForCall() with actual incoming SIP
    calls or PRESENTATION_RESTRICTED calls, but I'll ask QA to do that.

Bug: 4998476

Change-Id: I31ec7606bf47cbf4531b56fa925a097c34dad912