In-call UI should put the system bar in 'lights out' mode
Since we want this for the entire activity (not just for a specific view),
we do this by setting the SYSTEM_UI_FLAG_LOW_PROFILE flag in our window's
params.systemUiVisibility field.

Also, a minor followup to change I02140881 (bug 5161286), which disabled
navigation in the system bar while an incoming call is ringing.  I
originally implemented that using the StatusBarManager DISABLE_NAVIGATION
flag.  But the DISABLE_NAVIGATION flag only affects the "Home" and
"Recents" buttons, so I actually should set the DISABLE_BACK flag too.
(This also has the nice side-effect that the system bar will now
*completely* disappear while an incoming call is ringing.)

Tested (on a non-capacitive-soft-buttons device):
(1) In-call UI for an outgoing call
      ==> The system bar now shows little dots instead of the usual
           Back/Home/Recents icons.
(2) Incoming-call UI
      ==> The system bar is now totally hidden (and inactive).
(3) In-call UI after answering an incoming call:
      ==> Same behavior as (1).

Bug: 5252945

Change-Id: I971678ac565aa16d5e5b4c024fcf504ed2d74017