Temporarily disable android:animateLayoutChanges
Temporarily remove android:animateLayoutChanges="true" from the two places
in the in-call view hierarchy that use it.  This is a (temporary)
workaround for bug 5262542, which was causing the main in-call  button
cluster to not properly appear after answering an incoming call.

TESTED: On crespo, answer an incoming call.  Without this change, in-call
buttons are missing (100% reproduceable.)  With this change, in-call
buttons appear correctly.

Bug: 5262542
Change-Id: I95eda97e7dc88f312d42a3e0f5df97bb36115667