Enable in-call "End" button even if phone is idle
We've gotten occasional reports of touch input problems leaving you
"stuck" on the in-call screen.  By keeping the "End" button always
enabled, we'll now be able to tell from the logs whether touch input is
totally broken, or if instead we just got stuck on the InCallScreen even
though the phone was idle.

We might *also* want to bail out immediately in this case (to make it
easier for the user to recover if we *do* get stuck on the InCallScreen
with the phone IDLE), but I'm not enabling that yet since it might mask
some other bug.  I added a TODO comment about that, though.

Bug: 5331988
Change-Id: I1a297200898f9c1477bd556b2107c80e52b542e5