OTASP: Fix problems with how we handle the PERFORM_CDMA_PROVISIONING intent
When the SetupWizard notices that the device needs CDMA activation, it
fires off a PERFORM_CDMA_PROVISIONING intent.

Prior to this change, we handled that intent by simply launching a
call to "*228", which had two problems:
  (1) We're not supposed to launch the call right away; we're actually
      supposed to let the user press the "Activate" button first
  (2) Once the call would start, the onscreen OTASP UI would be in the
      wrong state: the DTMF dialpad would be hidden, even though the
      activation process needs you to press "1" :-(

Now, we handle the PERFORM_CDMA_PROVISIONING intent by simply bringing
up the InCallScreen in the "activation" state.  Then, when the user
presses the "Activate" button we'll actually launch the call.

TESTED: by manually firing off a PERFORM_CDMA_PROVISIONING intent:
    adb shell am start -a com.android.phone.PERFORM_CDMA_PROVISIONING
Result:
  - The "Activate your phone" screen came up correctly
  - When I pressed "Activate", it successfully launched the OTASP call
  - Once the call started, the DTMF dialpad *was* displayed correctly, and I
    was able to press "1" to activate my device.
  (The activation ultimately failed, though, since my device was already
  activated.)

Bug: 5092031

Change-Id: If3166139ea162dca2ff6d4dae52345bcd4414faa