(Re)enable layout transition animations on the in-call screen
Explicitly fade-in and fade-out a couple of in-call UI elements:
  - the call state label that briefly shows "Dialing"
  - the in-call DTMF dialpad
to make state transitions look smoother.

Earlier in ICS we did this using android:animateLayoutChanges="true" on
these views' parent containers, but that had some undesirable side effects
(see bug 5315523.)  Now, we just manually fade in/out the specific
elements we want to animate, and keep the code simple by using a
convenient utility class from Chet (see his blog post
http://graphics-geek.blogspot.com/2011/09/old-views-dont-die-they-just-fade-away.html.)

TESTED:
  Call state label:
    - visible instantly at the start of a call
    - visible instantly when "Call ended" state starts
    - fades out smoothly after "Dialing" state ends
  DTMF dialpad:
    - nice crossfade with the regular CallCard
    - No unexpected fading when *starting* a call, even if the dialpad was
      up when the previous call ended
    - Bang as fast as you can on dialpad button: always ends up in a sane
      state with either dialpad or CallCard visible (but not both, or neither)

Note there are a couple of other state transitions we'd like to improve as
well, but those are GSM-specific ("Manage conference" and the "two lines
in use" state) so they can wait till MR1.

Bug: 5315523

Change-Id: I52d060d7ca66ac936a6daffb297bf0e8d80f6848