Force in-call notification to be shown if "voice privacy" is active
Normally we suppress the in-call status bar notification if the
InCallScreen is the foreground activity.  (The goal is to reduce clutter
in the status bar; that icon isn't needed since it's already obvious that
you're on a call.)

But we shouldn't do that if "Voice Privacy" mode is active, since the
status bar icon is the only indication we have for "voice privacy" mode.

So if voice privacy is in effect, force the status bar icon to be visible
regardless of whether the InCallScreen is active.

TESTED:

- NOTE I couldn't do a real end-to-end test because the "voice privacy"
  feature isn't available at all on my device for some reason.  So
  instead, I tested this change by simply hacking getVoicePrivacyState()
  to return true.

(1) Made a call with Voice Privacy "enabled":
  ==> Status bar icon showing phone with tiny lock symbol is visible
      when I'm on the InCallScreen *and* when I bail out to some other app

(2) Made a call  with Voice Privacy disabled:
  ==> Status bar icon is visible *only* after I bail out of the
      InCallScreen.  Also, the icon is just a plain phone, with no lock
      symbol.

Bug: 5371658

Change-Id: I230bed89eae137d4cd106f3e44d7ad0105d6998c