Fix the "Exiting Emergency Callback Mode" dialog
If you make an outgoing non-emergency call while in ECM, we're supposed to
(a) cancel ECM, and (b) display a simple dialog saying "Exiting Emergency
Callback Mode" when the in-call UI comes up.

This change fixes part (b), which was broken.  The problem was a bug in
the CallController: there's a "pending call status code" that tells the
InCallScreen to display the "Exiting ECM" dialog, we were setting it
correctly at the point we noticed we were in ECM, but we would then clear
it immediately as soon as the call succeeded :-(

This fix is for placeCallInternal() to explicitly return EXITED_ECM in
this case, rather than just plain SUCCESS.

TESTED:
  - Regular phone calls => No change
  - Exiting ECM => Dialog comes up correctly
  - Other types of call failures => No change (we still display the
    appropriate error dialogs).

Bug: 5487455
Change-Id: Ic5d521e5c95e8137067ca53b5876b35788768abc