Phone: Play an out of service tone if SIM card is removed during call
In the audio framework, if no tone is played at call disonnection, the output is not
released correctly in accordance to the platform audio drivers. Thus the audio is not
reset gracefully and the Modem ALSA driver not closed. As a result it is not possible
to playback a stream after a call ended due to a SIM removal. Playing an out of service
tone solves the problem.

Change-Id: I23a46951b903f9c01d3096a27dd80ff2f318187e
Author: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Beare, Bruce J <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 11908