Use isPotentialLocalEmergencyNumber() to enforce ACTION_CALL restriction
The framework PhoneNumberUtils class now provides a way to distinguish
between (a) numbers that are definitely emergency numbers, and (b) numbers
that *might* result in an emergency call being dialed, but aren't
specifically emergency numbers themselves.

Given that, we now use the new isPotentialLocalEmergencyNumber() API when
enforcing the restriction that 3rd party apps should not be allowed to
make emergency calls using the ACTION_CALL intent.  (This ensures that 3rd
party apps can't make emergency calls by passing in an "invalid" number
like "9111234" that isn't technically an emergency number but might still
result in an emergency call with some networks.)

Everywhere else in the app, though, we still use the original
isLocalEmergencyNumber() API, which now returns true only if the specified
number *exactly* matches a known emergency number.  This ensures that the
in-call UI will only display the "emergency call" state for numbers that are
*definitely* emergency numbers.  (See bug 5493790 for the full details.)

TESTED (on Prime-C):

- Call regular non-emergency numbers from the built-in dialer
  ==> calls succeed with no emergency-call specific behavior

- Call "911" from the built-in dialer
  ==> Call to 911 succeeds
  ==> In-call UI shows the "emergency call" state
  ==> We correctly disable "mute"
  ==> We correctly enter ECM

- Call "9111234567" from the built-in dialer
  ==> We allow the call to be placed
  ==> The call does not actually go through; you just hear a recorded
      message from Verizon saying "the call could not be completed"
      (although this behavior might be different on other networks)
  ==> The in-call UI does NOT show the "emergency call" state
  ==> We do not disable "mute"
  ==> We do not enter ECM

- use CallDialTest activity to fire off various ACTION_CALL intents
  (as if launched from a 3rd party app):
    "911"  ==> doesn't allow the call to be placed (brings up dialer instead)
    "9111234"  ==> doesn't allow the call to be placed (brings up dialer instead)
    "6502530000"  ==> call succeeds

Note: This change depends on
  Change-Id: Ic528cfcc555734cdaf4ca8a18a50199771ba49b1
in frameworks/base (which must be submitted first.)

Bug: 5493790
Change-Id: Ib949fea3c0ce6b341a90e617a03ba3f22c69018b