Fix potential crash when launching OTASP in response to BOOT_COMPLETED
When calling OtaUtils.maybeDoOtaCall() from the BOOT_COMPLETED receiver in
OtaStartupReceiver.java, we're running in the PhoneApp's context, not an
Activity context.  So if we do actually need to launch an OTASP call(*), the
FLAG_ACTIVITY_NEW_TASK flag is required.

(*) This issue first came up in the context of bug 5528222, although the
    real problem in *that* bug was that we shouldn't have been launching
    old-style OTASP at all on LTE devices.  But the fix here is still
    needed, for CDMA devices that do still use old-style OTASP.

TESTED:

(1) On Prime-C, temporarily hacked OtaUtils.maybeDoOtaCall() to force an
    old-style OTASP call upon boot.
    ==> Before this change: the startActivity() call crashes (as reported in
        bug 5528222)
    ==> After this change: old-style activation UI comes up correctly, and
        OTASP call starts when you press "Activate".

(2) Also confirmed that I didn't break the PERFORM_CDMA_PROVISIONING intent,
    by running
        adb shell am start -a com.android.phone.PERFORM_CDMA_PROVISIONING
    (which is the only other way to launch the old-style OTASP sequence.)
    That works fine both with and without this change.

Bug: 5528222
Change-Id: I89055c7b40a3e261aff3657a8ec052477462171e