Save network mode on GSM devices.
We save the mode in OnClick and handleGetPreferredNetworkTypeResponse
so the preferred network mode tracks the UI as best as possible.

On GSM devices when "Use only 2G networks" was
checed in Settings > More > Mobile Networks the
value wasn't saved. Thus when rebooting the mode
would revert to the previous value.

Looking at the history, it has been this way a
long time, but it must have worked because the
radio preserved the value, returning when the
current PREFERRED_NETWORK_TYPE was requested.

With ICS Samsung upgraded the Crespo radio and the value
doesn't always appear to be preserved. In any case
its better for the framework to preserve the value
rather than relying on actions of the many radio
implementations.

Bug: 5764203
Change-Id: I9ed5db4e5c2db01cdd54bc1cc36395326b209fe3