Remove BT specific icons for phone notifications
Phone app uses a call icon with Bluetooth symbol for its notifications,
but it is so small and thus not so informative.
Also because of bug 5439643 it isn't shown appropriately.

This change just removes the entire logic, showing a call icon without
the BT symbol.

Also this change removes stat_sys_phone_call_ringing, which are
identical to stat_sys_phone_call.

TESTED:
- Connect BT headset and receive a phone call. After that
  press home, exiting the in-call UI. See notification icon and see
  if BT symbol is gone.
- Try the sequence above without using BT headset.

Bug: 5439643
Change-Id: I3ae36b8e83009d2e39608d9ac7e3754dbd22f576