Rely on Java constructor to set default ringtone.
Ringer#mCustomRingtoneUri must not be null and CallNotifier
should initialize the Uri before calling Ringer#ring() with the
default Uri.

Bug 5527862 implies the default Uri may not be set in some rare case.
Relying on constructor should avoid the issue from the beginning.

Because the process won't be killed, we still need to reset the Uri
to default everytime.

TESTED:
- receive a phone call from a contact without a custom ringtone.
- receive a phone call from a contact with a custom ringtone.
(Those tests are for checking if the current feature is working fine.
 Bug 5527862 looks a very rare edge case and thus hard to test)

Bug: 5527862
Change-Id: I67c0bca9aa186a2986a91cc2baa5d7d57ce20b24