Stop updating the whole screen on onPause()
Fixing the issue 5967067: Avatar picture disappears too early on
hang-up.

On hang-up, Phone app first receives IDLE state and exits the screen.
The call state is IDLE on onPause() in that case, which will show brief
black screen reported in the issue.

Because we always refresh the screen on onResume(), we don't need to
refresh the whole screen on onPause().

TESTED:
- make a phone call, and hang it up
- make a phone call, and let the receiver hang up the call
- make a phone call, and bail out the screen. Then enter the in-call
  UI again. Try it during a conference call and manage-conference
  screen.

Bug: 5967067
Change-Id: Id354c1096a4ac30c4d398d3d169d2794060349d7