Avoid switch to landscape mode when initiating outbound call.
When holding your phone in landscape mode, both Dialtacts and the
in-call UI remain in portrait mode.  However, there are two activities
that occur before the in-call UI is entered (OutgoingCallBroadcaster
and SipCallOptionHandler).  These were allowing the screen orientation
to briefly switch to landscape, causing an ugly jank before switching
back to portrait mode for the in-call UI.

Fix this by forcing both activities to remain in portrait mode.

Bug: 6042351
Change-Id: I867fb314820a7419bab781936be4dbc0ded00d7a