Introduce fallback cache for ringtone/send-to-vm
Right now the app just relies on ContactsProvider for custom
ringtone and "send to voicemail" and falls back to system
default when it fails to obtain those info in 500msec, which often
happens when the system is overloaded or database is being synced.

This change introduces fallback cache for the latter case. The
cache is set up on the device boot, and is refreshed periodically.
It is used by CallNotifier only when the first lookup fails.

We do *not* fully rely on the local in-memory cache since it may
be stale at that moment, but we do not rely on ContentObserver
mechanism either to keep the cache super-fresh every time, because
it also will overload ContactsProvider, just for obtaining two tiny
info.

This change also prepares a BroadcastReceiver which is accessible
from Contacts editor screen. See "NEXT TODO" below.

NEXT TODO:
- Should have a mechanism to force update the cache from editor UI.
  Without that the cache is being obsolete for longer time.

TESTED:
- Set custom ringtone and/or send-to-voicemail flag and see the
  behavior.
- Insert intentional delay in async query code, forcing the fallback
  behavior.

Bug: 1735194
Bug: 5568640
Bug: 5718903
Bug: 6020594
Bug: 6078996
Change-Id: Ib9a592e0db69683dc08753e9fd617fe7f89c9e71