Fix obvious message leak.
Right now we're not removing the timeout message, so any kind of
outgoing call will reach the timeout. Though we won't show the
spinner thing itself, we emit mysterious log every time.

This change remove the unnecessary message when it becomes unnecessary.
It also delays the message until the broadcast being actually fired,
because the case is the only possible bottle neck in the sequence.
Other cases will be handled in UI thread so we won't show the spinner
anyway.

Change-Id: Ic58928b8cf256d080f2a8d9b3f56fa85a5b9e62d