Stop relying on person_id on notification photo
Fixing bug 6258768 : Default avatar is not displayed in the notification
tray while call is in progress.

It happens when there's no photo found for the contact. At that moment
the NotificationMgr tries to use Uri with person_id, which fails because
there's no photo anyway. We probably want to show some default icon
there, while at that moment Notification is already fired.

To conditionally show photo or other default icons, it takes too much work
with less fruit; we need to setup async task and send notification
afterward. Note that if ContactsProvider can return the photo, it should
have been done during or even before RINGING state in most of the cases.

Also, eventually we will need to rely on Notification.Builder to fix
bug 5161272, which just supports Bitmap for large icon; we cannot rely
on Uri nor resource id.

So this change just relys on the cases which are able to use Bitmap for
icons. For other cases (which should be hopefully rare) we just use same
default icons as before.

TESTED:
- make/receive a phone call from a contact with photo. Users should be
  able to see the contact photo on ongoing call notification.

- make/receive a phone call from a contact without photo. Users should
  be able to see default icons.

- receive a phone call, answer it, and press Home immediately. Users can
  briefly see remaining "incoming call" notification (not ongoing call
  notification). At this point CallerInfo doesn't have correct photo
  even when it will be available eventually, so users should be able to
  see default icons at least.

Bug: 6258768
Change-Id: Ifdb5c6782dcfbe24563a7231b4adb4371a4e8c53